added compatibility for python 3 and added refractive index calculation for C band for selected binaries

+++ src/openbandparams/examples/Plot_Bandgap_vs_Lattice_Constant_of_Quaternary3.py
@@ -70,10 +70,10 @@ for x, y, label in zip(xs, y_L, labels):
 # plot the quaternary
 indices = numpy.arange(100)
 fractions = numpy.linspace(0, 1, 100)
-x = numpy.empty(100, dtype=numpy.float)
-y_Gamma = numpy.empty(100, dtype=numpy.float)
-y_X = numpy.empty(100, dtype=numpy.float)
-y_L = numpy.empty(100, dtype=numpy.float)
+x = numpy.empty(100, dtype=numpy.float32)
+y_Gamma = numpy.empty(100, dtype=numpy.float32)
+y_X = numpy.empty(100, dtype=numpy.float32)
+y_L = numpy.empty(100, dtype=numpy.float32)
 first = True
 for xfrac in numpy.linspace(0, 1, 10):
     for i, yfrac in zip(indices, fractions):
